Moved slow test to FuncTest
