+++ NuGet.Core.sln
@@ -125,6 +125,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "funcTest", "funcTest", "{BB
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NuGet.Commands.FuncTest", "test\NuGet.Core.FuncTests\NuGet.Commands.FuncTest\NuGet.Commands.FuncTest.xproj", "{63C0BFF6-508D-4E5F-BBBC-A636D1398C6B}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NuGet.Protocol.FuncTest", "test\NuGet.Core.FuncTests\NuGet.Protocol.FuncTest\NuGet.Protocol.FuncTest.xproj", "{1A94F70B-1AB6-4638-931A-4AB0EBEF7EC1}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -351,6 +353,10 @@ Global
 		{63C0BFF6-508D-4E5F-BBBC-A636D1398C6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{63C0BFF6-508D-4E5F-BBBC-A636D1398C6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{63C0BFF6-508D-4E5F-BBBC-A636D1398C6B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1A94F70B-1AB6-4638-931A-4AB0EBEF7EC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1A94F70B-1AB6-4638-931A-4AB0EBEF7EC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1A94F70B-1AB6-4638-931A-4AB0EBEF7EC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1A94F70B-1AB6-4638-931A-4AB0EBEF7EC1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -410,5 +416,6 @@ Global
 		{7FACD9D7-F70D-41C4-BE61-2753E018057D} = {7323F93B-D141-4001-BB9E-7AF14031143E}
 		{D8E35E61-07AE-41D2-BB40-F3BDD8770A92} = {7323F93B-D141-4001-BB9E-7AF14031143E}
 		{63C0BFF6-508D-4E5F-BBBC-A636D1398C6B} = {BB63CACA-866F-454F-BA4C-78B788D032BB}
+		{1A94F70B-1AB6-4638-931A-4AB0EBEF7EC1} = {BB63CACA-866F-454F-BA4C-78B788D032BB}
 	EndGlobalSection
 EndGlobal
